+++ allsecureexchange/upload/catalog/language/en-gb/extension/payment/allsecure_exchange.php
@@ -1,6 +1,7 @@
 <?php
 $_['order_cancelled'] = 'Order was canceled.';
 $_['order_error'] = 'An error occurred during the payment process. Please try again.';
+$_['order_on_hold'] = 'Awaiting capture/void.';
 $_['button_confirm'] = 'Confirm Order';
 
 $_['redirect_text'] = 'You are being redirected. Please wait.';