+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## Unreleased
+### Added
+- Preauthorize/Capture/Void
+### Changed
+- Order ID suffix to prevent TX ID conflicts
+
 ## v1.2.2 (2019-12-10)
 ### Fixed
 - Gateway client PHP 7.3 compatibility: remove redundant filter_var FILTER_VALIDATE_URL flags